#8 移動処理の仕様変更．1マスずつ移動する．ユニットの位置をクリック又は移動コストが0の場合に移動終了．例外処理の追加．

+++ Assets/Scenes/Scripts/Unit/UnitStatus.cs
@@ -4,10 +4,10 @@ using System.Collections.Generic;
 public class UnitStatus : MonoBehaviour
 {
     //操作ユニットの基本ステータス
-    public float HitPoint;
-    public float AttackPoint;
-    public float DefensPoint;
-    public float MovementPoint;
+    public int HitPoint;
+    public int AttackPoint;
+    public int DefensPoint;
+    public int MovementPoint;
 
     void Start()
     {
